+++ Ship.cpp
@@ -1,11 +1,11 @@
 #include "Ship.h"
 
-Ship::Ship(std::vector<Brick*>& bricksForShip, int size)
+Ship::Ship(std::vector<Brick*>& bricksForShip, size_t size)
 {
 	this->size = size;
 	shipsBricks = bricksForShip;
 
-	for (int i = 0; i < size; i++)
+	for (size_t i = 0; i < size; i++)
 	{
 		shipsBricks[i]->SetPartOfAShip();
 	}
